Abstract DatabaseInfoCache/TableIDMap
