+++ dbms/src/TiDB/Schema/TiDBSchemaSyncer.h
@@ -18,7 +18,9 @@
 #include <Common/Stopwatch.h>
 #include <Debug/MockSchemaGetter.h>
 #include <Debug/MockSchemaNameMapper.h>
+#include <TiDB/Schema/DatabaseInfoCache.h>
 #include <TiDB/Schema/SchemaBuilder.h>
+#include <TiDB/Schema/TableIDMap.h>
 #include <TiDB/Schema/TiDB.h>
 #include <pingcap/kv/Cluster.h>
 #include <pingcap/kv/Snapshot.h>
@@ -51,29 +53,23 @@ private:
     // for syncSchemas
     std::mutex mutex_for_sync_schema;
 
-    // mutex for databases
-    std::shared_mutex shared_mutex_for_databases;
-
-    std::unordered_map<DB::DatabaseID, TiDB::DBInfoPtr> databases;
-
     LoggerPtr log;
 
+    DatabaseInfoCache databases;
     TableIDMap table_id_map;
 
     Getter createSchemaGetter(KeyspaceID keyspace_id)
     {
-        [[maybe_unused]] auto tso = cluster->pd_client->getTS();
         if constexpr (mock_getter)
         {
             return Getter();
         }
         else
         {
+            auto tso = cluster->pd_client->getTS();
             return Getter(cluster.get(), tso, keyspace_id);
         }
     }
-
-    std::tuple<bool, DatabaseID, TableID> findDatabaseIDAndTableID(TableID physical_table_id);
 
 public:
     TiDBSchemaSyncer(KVClusterPtr cluster_, KeyspaceID keyspace_id_)
@@ -114,32 +110,13 @@ private:
 
     TiDB::DBInfoPtr getDBInfoByName(const String & database_name) override
     {
-        std::shared_lock<std::shared_mutex> lock(shared_mutex_for_databases);
-
-        auto it = std::find_if(databases.begin(), databases.end(), [&](const auto & pair) {
-            return pair.second->name == database_name;
-        });
-        if (it == databases.end())
-            return nullptr;
-        return it->second;
-    }
-
-    TiDB::DBInfoPtr getDBInfoByMappedName(const String & mapped_database_name) override
-    {
-        std::shared_lock<std::shared_mutex> lock(shared_mutex_for_databases);
-
-        auto it = std::find_if(databases.begin(), databases.end(), [&](const auto & pair) {
-            return NameMapper().mapDatabaseName(*pair.second) == mapped_database_name;
-        });
-        if (it == databases.end())
-            return nullptr;
-        return it->second;
+        return databases.getDBInfoByName(database_name);
     }
 
     void dropAllSchema(Context & context) override
     {
         auto getter = createSchemaGetter(keyspace_id);
-        SchemaBuilder<Getter, NameMapper> builder(getter, context, databases, table_id_map, shared_mutex_for_databases);
+        SchemaBuilder<Getter, NameMapper> builder(getter, context, databases, table_id_map);
         builder.dropAllSchema();
     }
 
@@ -147,11 +124,7 @@ private:
     // just for testing restart
     void reset() override
     {
-        {
-            std::unique_lock<std::shared_mutex> lock(shared_mutex_for_databases);
-            databases.clear();
-        }
-
+        databases.clear();
         table_id_map.clear();
         cur_version = 0;
     }
